实现 Token 鉴权

+++ online-judge-backend/src/Control.cpp
@@ -13,6 +13,9 @@
 using namespace std;
 
 // ++++++++++++++++++++++++++++++ 用户模块 Start ++++++++++++++++++++++++++++++
+// 无权限（该变量用于当无权限情况时，直接返回该变量）
+static Json::Value NoPermission;
+
 // 注册用户
 Json::Value Control::RegisterUser(Json::Value &registerjson)
 {
@@ -34,24 +37,40 @@ Json::Value Control::SelectUserInfo(Json::Value &queryjson)
 // 更新用户信息
 Json::Value Control::UpdateUserInfo(Json::Value &updatejson)
 {
+    // 如果不是本人或者管理员，无权修改
+    if (!UserList::GetInstance()->IsAuthor(updatejson))
+        return NoPermission;
+
     return UserList::GetInstance()->UpdateUserInfo(updatejson);
 }
 
 // 获取用户信息以提供修改
 Json::Value Control::SelectUserUpdateInfo(Json::Value &queryjson)
 {
+    // 如果不是本人或者管理员，无权修改
+    if (!UserList::GetInstance()->IsAuthor(queryjson))
+        return NoPermission;
+
     return UserList::GetInstance()->SelectUserUpdateInfo(queryjson);
 }
 
 // 管理员分页获取用户信息
 Json::Value Control::SelectUserSetInfo(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return UserList::GetInstance()->SelectUserSetInfo(queryjson);
 }
 
-// 删除用户
+// 管理员删除用户
 Json::Value Control::DeleteUser(Json::Value &deletejson)
 {
+    // 如果不是管理员，无权删除
+    if (!UserList::GetInstance()->IsAdministrator(deletejson))
+        return NoPermission;
+
     return UserList::GetInstance()->DeleteUser(deletejson);
 }
 
@@ -60,12 +79,22 @@ Json::Value Control::SelectUserRank(Json::Value &queryjson)
 {
     return UserList::GetInstance()->SelectUserRank(queryjson);
 }
+
+// 登录用户通过Token（Token 鉴权实现）
+Json::Value Control::LoginUserByToken(Json::Value &loginjson)
+{
+    return UserList::GetInstance()->LoginUserByToken(loginjson);
+}
 // ++++++++++++++++++++++++++++++ 用户模块 End ++++++++++++++++++++++++++++++
 
 // ++++++++++++++++++++++++++++++ 题目模块 Start ++++++++++++++++++++++++++++++
 // 管理员查看题目详细信息
 Json::Value Control::SelectProblemInfoByAdmin(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return ProblemList::GetInstance()->SelectProblemInfoByAdmin(queryjson);
 }
 
@@ -75,9 +104,13 @@ Json::Value Control::SelectProblem(Json::Value &queryjson)
     return ProblemList::GetInstance()->SelectProblem(queryjson);
 }
 
-// 编辑题目
+// 管理员编辑题目
 Json::Value Control::EditProblem(Json::Value &insertjson)
 {
+    // 如果不是管理员，无权编辑
+    if (!UserList::GetInstance()->IsAdministrator(insertjson))
+        return NoPermission;
+
     Json::Value resjson;
     if (insertjson["EditType"].asString() == "Insert")
     {
@@ -91,9 +124,13 @@ Json::Value Control::EditProblem(Json::Value &insertjson)
     return resjson;
 }
 
-// 删除题目
+// 管理员删除题目
 Json::Value Control::DeleteProblem(Json::Value &deletejson)
 {
+    // 如果不是管理员，无权删除
+    if (!UserList::GetInstance()->IsAdministrator(deletejson))
+        return NoPermission;
+
     return ProblemList::GetInstance()->DeleteProblem(deletejson);
 }
 
@@ -103,9 +140,13 @@ Json::Value Control::SelectProblemList(Json::Value &queryjson)
     return ProblemList::GetInstance()->SelectProblemList(queryjson);
 }
 
-// 管理员查询列表
+// 管理员查询题目列表
 Json::Value Control::SelectProblemListByAdmin(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return ProblemList::GetInstance()->SelectProblemListByAdmin(queryjson);
 }
 
@@ -123,9 +164,13 @@ Json::Value Control::GetTags(Json::Value &queryjson)
 // ++++++++++++++++++++++++++++++ 题目模块 End ++++++++++++++++++++++++++++++
 
 // ++++++++++++++++++++++++++++++ 公告模块 Start ++++++++++++++++++++++++++++++
-// 添加公告
+// 管理员添加公告
 Json::Value Control::InsertAnnouncement(Json::Value &insertjson)
 {
+    // 如果不是管理员，无权添加
+    if (!UserList::GetInstance()->IsAdministrator(insertjson))
+        return NoPermission;
+
     return AnnouncementList::GetInstance()->InsertAnnouncement(insertjson);
 }
 
@@ -138,6 +183,10 @@ Json::Value Control::SelectAnnouncementList(Json::Value &queryjson)
 // 管理员分页查询公告
 Json::Value Control::SelectAnnouncementListByAdmin(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return AnnouncementList::GetInstance()->SelectAnnouncementListByAdmin(queryjson);
 }
 
@@ -147,21 +196,33 @@ Json::Value Control::SelectAnnouncement(Json::Value &queryjson)
     return AnnouncementList::GetInstance()->SelectAnnouncement(queryjson);
 }
 
-// 查询公告 进行编辑
+// 管理员查询公告 进行编辑
 Json::Value Control::SelectAnnouncementByEdit(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看并编辑
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return AnnouncementList::GetInstance()->SelectAnnouncementByEdit(queryjson);
 }
 
-// 更新公告
+// 管理员更新公告
 Json::Value Control::UpdateAnnouncement(Json::Value &updatejson)
 {
+    // 如果不是管理员，无权修改
+    if (!UserList::GetInstance()->IsAdministrator(updatejson))
+        return NoPermission;
+
     return AnnouncementList::GetInstance()->UpdateAnnouncement(updatejson);
 }
 
-// 删除公告
+// 管理员删除公告
 Json::Value Control::DeleteAnnouncement(Json::Value &deletejson)
 {
+    // 如果不是管理员，无权删除
+    if (!UserList::GetInstance()->IsAdministrator(deletejson))
+        return NoPermission;
+
     Json::Value resjson = AnnouncementList::GetInstance()->DeleteAnnouncement(deletejson);
 
     // 当评论模块完成时，将下面注释去掉
@@ -176,9 +237,13 @@ Json::Value Control::DeleteAnnouncement(Json::Value &deletejson)
 // ++++++++++++++++++++++++++++++ 公告模块 End ++++++++++++++++++++++++++++++
 
 // ++++++++++++++++++++++++++++++ 讨论模块 Start ++++++++++++++++++++++++++++++
-// 添加讨论
+// 用户添加讨论
 Json::Value Control::InsertDiscuss(Json::Value &insertjson)
 {
+    // 如果不是用户，无权添加
+    if (!UserList::GetInstance()->IsOrdinaryUser(insertjson))
+        return NoPermission;
+
     return DiscussList::GetInstance()->InsertDiscuss(insertjson);
 }
 
@@ -191,6 +256,10 @@ Json::Value Control::SelectDiscussList(Json::Value &queryjson)
 // 管理员分页查询讨论
 Json::Value Control::SelectDiscussListByAdmin(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return DiscussList::GetInstance()->SelectDiscussListByAdmin(queryjson);
 }
 
@@ -203,18 +272,30 @@ Json::Value Control::SelectDiscuss(Json::Value &queryjson)
 // 查询讨论的详细信息，主要是编辑时的查询
 Json::Value Control::SelectDiscussByEdit(Json::Value &queryjson)
 {
+    // 如果不是用户，无权查看
+    if (!UserList::GetInstance()->IsOrdinaryUser(queryjson))
+        return NoPermission;
+
     return DiscussList::GetInstance()->SelectDiscussByEdit(queryjson);
 }
 
-// 修改评论数的数量
+// 更新讨论
 Json::Value Control::UpdateDiscuss(Json::Value &updatejson)
 {
+    // 如果不是本人或者管理员，无权修改
+    if (!UserList::GetInstance()->IsAuthor(updatejson))
+        return NoPermission;
+
     return DiscussList::GetInstance()->UpdateDiscuss(updatejson);
 }
 
 // 删除讨论
 Json::Value Control::DeleteDiscuss(Json::Value &deletejson)
 {
+    // 如果不是本人或者管理员，无权删除
+    if (!UserList::GetInstance()->IsAuthor(deletejson))
+        return NoPermission;
+
     Json::Value resjson = DiscussList::GetInstance()->DeleteDiscuss(deletejson);
 
     // 当评论模块完成时，将下面注释去掉
@@ -238,6 +319,10 @@ Json::Value Control::SelectSolutionList(Json::Value &queryjson)
 // 管理员查询
 Json::Value Control::SelectSolutionListByAdmin(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return SolutionList::GetInstance()->SelectSolutionListByAdmin(queryjson);
 }
 
@@ -250,24 +335,40 @@ Json::Value Control::SelectSolution(Json::Value &queryjson)
 // 查询题解进行编辑
 Json::Value Control::SelectSolutionByEdit(Json::Value &queryjson)
 {
+    // 如果不是用户，无权查看
+    if (!UserList::GetInstance()->IsOrdinaryUser(queryjson))
+        return NoPermission;
+
     return SolutionList::GetInstance()->SelectSolutionByEdit(queryjson);
 }
 
 // 插入题解
 Json::Value Control::InsertSolution(Json::Value &insertjson)
 {
+    // 如果不是用户，无权添加
+    if (!UserList::GetInstance()->IsOrdinaryUser(insertjson))
+        return NoPermission;
+
     return SolutionList::GetInstance()->InsertSolution(insertjson);
 }
 
 // 更新题解
 Json::Value Control::UpdateSolution(Json::Value &updatejson)
 {
+    // 如果不是本人或者管理员，无权修改
+    if (!UserList::GetInstance()->IsAuthor(updatejson))
+        return NoPermission;
+
     return SolutionList::GetInstance()->UpdateSolution(updatejson);
 }
 
 // 删除题解
 Json::Value Control::DeleteSolution(Json::Value &deletejson)
 {
+    // 如果不是本人或者管理员，无权删除
+    if (!UserList::GetInstance()->IsAuthor(deletejson))
+        return NoPermission;
+
     Json::Value resjson = SolutionList::GetInstance()->DeleteSolution(deletejson);
 
     // 当评论模块完成时，将下面注释去掉
@@ -285,6 +386,10 @@ Json::Value Control::DeleteSolution(Json::Value &deletejson)
 // 管理员查询评论
 Json::Value Control::SelectCommentListByAdmin(Json::Value &queryjson)
 {
+    // 如果不是管理员，无权查看
+    if (!UserList::GetInstance()->IsAdministrator(queryjson))
+        return NoPermission;
+
     return CommentList::GetInstance()->SelectCommentListByAdmin(queryjson);
 }
 
@@ -335,6 +440,10 @@ Json::Value Control::InsertComment(Json::Value &insertjson)
 // 删除评论
 Json::Value Control::DeleteComment(Json::Value &deletejson)
 {
+    // 如果不是用户，没有权限删除
+    if (!UserList::GetInstance()->IsOrdinaryUser(deletejson))
+        return NoPermission;
+
     string articleid = deletejson["ArticleId"].asString();
 
     Json::Value resjson;
@@ -390,6 +499,10 @@ Json::Value Control::SelectStatusRecord(Json::Value &queryjson)
 // ++++++++++++++++++++++++++++++ 判题模块 Start ++++++++++++++++++++++++++++++
 Json::Value Control::GetJudgeCode(Json::Value judgejson)
 {
+    // 如果不是用户，无权判题
+    if (!UserList::GetInstance()->IsOrdinaryUser(judgejson))
+        return NoPermission;
+
     Json::Value resjson;
     // 传入Json(ProblemId, UserId, UserNickName, Code, Language, TimeLimit, MemoryLimit, JudgeNum, ProblemTitle)
 
@@ -429,8 +542,8 @@ Json::Value Control::GetJudgeCode(Json::Value judgejson)
 
     // 更新状态信息
     /*
-        传入：Json(SubmitId,Status,RunTime,RunMemory,Length,ComplierInfo,
-        TestInfo[(Status,StandardOutput,PersonalOutput,RunTime,RunMemory)])
+        传入：Json(SubmitId, Status, RunTime, RunMemory, Length, ComplierInfo,
+        TestInfo[(Status, StandardOutput, PersonalOutput, RunTime, RunMemory)])
     */
     StatusRecordList::GetInstance()->UpdateStatusRecord(json);
     // 更新题目的状态
@@ -458,6 +571,13 @@ Control::Control()
 {
     // 初始化题目标签
     Tag::GetInstance()->InitProblemTags();
+
+    // 初始化用户权限
+    UserList::GetInstance()->InitUserAuthority();
+
+    // 初始化返回变量
+    NoPermission["Result"] = "401";
+    NoPermission["Reason"] = "无权限";
 }
 
 Control::~Control()
